currently we are able to login via twitter and niuu api
the api calls are going through kirks proxy,
the json is being passed as a string via the standard param data in the request.

this way the standard $_POST data is available on the receiving end.

the proxy on kirks server gives proper CORS security responses to the browser, to avoid around javascript security blocking

Signed-off-by: Kirk Peterkin <kirk@niiu.de>