loading the Facebook js sdk
taking a step away from Node and passport it is rather straightforward to load a page with the Facebook js sdk, and then allow users to click a button to authorize your app, then you have access to all of the data the requested.  in the dev folder is a folder called plunker which loads a page with an authorization via Facebook.
if I use the Facebook authorization button it seems to work fine but when i try to authorize with a function on click of another element the login function seems to get called immediately and not on click.

Signed-off-by: Kirk Peterkin <kirk@niiu.de>