YDN Safari Issues
So for some reason the YDN that I was using wasn't working as anticipated on safari.  We learned earlier that we had to define the keys differently so they would be accessible in safarii, but you can only have one key in a table which doesn't jive with this user info table Ive been using.

I think what I am going to do instead is to put this entire user info object as an object into safari and we will take it from there.

Signed-off-by: Kirk Peterkin <kirk@niiu.de>