good till refresh
at this point we are saving the items in the db in both browsers and we can pull them back out in both browsers as long as we hold onto the db object. I think when the page is refreshed and this object is lost from scope though we don't seem to be able to get it back in safari.  Ive tried using sql and a couple other methods i see documented on yen and they work fine from the original db object but not from a new one. in chrome they work fine either way

Signed-off-by: Kirk Peterkin <kirk@niiu.de>