YDN DB working in safari!!  Now able to retrieve user (or any other) information from the DB after the session is over or reset in safari as well as chrome.
ok so the problem was that if you lose the handle to the db.  you can recreate it at any time, but just creating a new object pointing to the same db.  The difference though is that in Safari, that object uses its own primary key, unless you initiate the object along with a schema, that tells it which part of the object you want to use as a key. In Chrome you only need to do this when you create the DB in safari you need to do it anytime you want to get a reference to the DB.

The /todo/todo.html file was an example from the YDN site that was able to use the keys to access the table even once the page had been refreshed.  I used it to recreate the behavior desired.

Signed-off-by: Kirk Peterkin <kirk@niiu.de>