getting articles on userhome from db
not sure why this wasn't working exactly but I was getting an error about a duplicate index in the repeater. angular makes a pseudo id when it ng-repeats through an array, but you can override this and tell it which property of your array should be used as a key using hg-repeat track by (here in user home).

Signed-off-by: Kirk Peterkin <kirk@niiu.de>