rearranged the articles in the userHome view by published date
took waaaay to long. the code for it is in here.  you need to use a track by clause in your repeater so it doesn't choke on duplicates and  the trck by needs to stay at the end, and the orderBy needs to have article.id rather than just id for some reason.

Signed-off-by: Kirk Peterkin <kirk@niiu.de>