Creating a 3s MenuObject via JS
previously we were building the 3s menu via angular filters. now we are building an array for it in the niuusyncer service. It should perhaps get moved onto its own somewhere.  At this point the source array that is created in a section is keyed to the sourceSubsection Id. I was thinking to make these just standard list numbers and keep the source_section ids elsewhere.

Signed-off-by: Kirk Peterkin <kirk@niiu.de>