dealing with arrays
so I was mapping the section id's to the array values so I could call the sectionObj I needed without looping through the array, but if you add say SectionObj[10] to an array angular assumes the array has 9 objects before it.   which leads to it creating a bunch of empty elements and then breaking in the repeater because the null elements are duplicates of one another. Here I am using the getByProperty filter that I built previously to get the sectionObjects by their id property.

Signed-off-by: Kirk Peterkin <kirk@niiu.de>