translation library
library defaults to german, but if the user has their roster language as english then the site is rendered in English as well. There is also a language button in the menu which switches the language when clicked. The changed language is stored in local storage, can also be in a cookie or the DB. Db probably makes the most sense since it is should change with each user.

Signed-off-by: Kirk Peterkin <kirk@niiu.de>