loading your articles on initial page without transitions, article content issue
previously you could refresh and not see article content.  I repaired that earlier but the article source logo was not being set in those cases. So now I have the article pull the source logo in the article service as well rather than looking in the article object for it. Maybe its best if this is done at the same time when the articles are initiated. either way its working for now.

Signed-off-by: Kirk Peterkin <kirk@niiu.de>